iot/battle-mlx-cam/requirements.txt
opencv-python>=4.8.0
requests>=2.31.0
python-dotenv>=1.0.0
websocket-client>=1.6.0
Pillow>=10.0.0
pyobjc-framework-Vision>=10.0
pyobjc-framework-Quartz>=10.0
customtkinter>=5.2.0
rembg>=2.0.50
onnxruntime>=1.15.0
